@@ -48,11 +48,10 @@ class ConfirmationPage extends React.Component {
 
     return (
       <div>
-        <h3 className="confirmation-page-title">Claim received</h3>
-        <p>We process claims in the order we receive them.</p>
+        <h3 className="confirmation-page-title">Claim submitted</h3>
+        <p>We process claims in the order we receive them. Please print this page for your records. We'll also send you an email to confirm that we've received your application. Please note, it may take up to 24 hours to receive this email.</p>
         <p>
-          We may contact you for more information or documents.<br/>
-          <i>Please print this page for your records.</i>
+          We may contact you for more information or documents.
         </p>
         <div className="inset">
           <h4>Burial Benefit Claim <span className="additional">(Form 21P-530)</span></h4>
@@ -64,7 +63,7 @@ class ConfirmationPage extends React.Component {
               <span>{response.confirmationNumber}</span>
             </li>
             <li>
-              <strong>Date received</strong><br/>
+              <strong>Date submitted</strong><br/>
               <span>{submittedAt.utcOffset(offset).format('MMM D, YYYY h:mm a [CT]')}</span>
             </li>
             <li>
