@@ -10,15 +10,13 @@ const AlertsList = ({ alerts, appealIsActive }) => {
 
   const allAlertsContent = alerts.map((alert) => getAlertContent(alert, appealIsActive));
 
-  const takeActionAlerts = allAlertsContent.filter((alert) => (alert.displayType === 'take_action'));
-  const infoAlert = allAlertsContent.filter((alert) => (alert.displayType === 'info'));
+  const takeActionAlertsPresent = allAlertsContent.some((alert) => (alert.displayType === 'take_action'));
 
-  const takeActionHeader = (takeActionAlerts.length)
+  const takeActionHeader = takeActionAlertsPresent
     ? (<h3>Take Action</h3>)
     : null;
 
-  const alertsList = takeActionAlerts
-    .concat(infoAlert)
+  const alertsList = allAlertsContent
     .map((alert, index) => {
       const key = `${alert.type}-${index}`;
       return (<Alert
