@@ -7,8 +7,8 @@ const AppealsV2TabNav = ({ appealId }) => {
   return (
     <div>
       <ul className="va-tabs claims-status-tabs small-12 medium-10" role="tablist">
-        <TabItem shortcut={1} className="appeals-tabs-item" id="v2status" tabpath={`appeals-v2/${appealId}/status`} title="Status"/>
-        <TabItem shortcut={2} className="appeals-tabs-item" id="v2detail" tabpath={`appeals-v2/${appealId}/detail`} title="Issues"/>
+        <TabItem shortcut={1} className="appeals-tabs-item" id="v2status" tabpath={`appeals/${appealId}/status`} title="Status"/>
+        <TabItem shortcut={2} className="appeals-tabs-item" id="v2detail" tabpath={`appeals/${appealId}/detail`} title="Issues"/>
       </ul>
     </div>
   );
