@@ -1,56 +1,6 @@
 import environment from './environment.js';
-import { updateLoggedInStatus, updateLogInUrls, FETCH_LOGIN_URLS_FAILED } from '../../login/actions';
+import { updateLoggedInStatus } from '../../login/actions';
 import { updateProfileFields, profileLoadingFinished } from '../../user-profile/actions';
-
-export function handleVerify(verifyUrl) {
-  window.dataLayer.push({ event: 'verify-link-clicked' });
-  if (verifyUrl) {
-    window.dataLayer.push({ event: 'verify-link-opened' });
-    const receiver = window.open(`${verifyUrl}&op=signin`, 'signinPopup', 'resizable=yes,scrollbars=1,top=50,left=500,width=500,height=750');
-    receiver.focus();
-  }
-}
-
-export function handleMultifactor(multifactorUrl) {
-  window.dataLayer.push({ event: 'multifactor-link-clicked' });
-  if (multifactorUrl) {
-    window.dataLayer.push({ event: 'multifactor-link-opened' });
-    const receiver = window.open(multifactorUrl, 'signinPopup', 'resizable=yes,scrollbars=1,top=50,left=500,width=500,height=750');
-    receiver.focus();
-  }
-}
-
-export function getVerifyUrl(onUpdateVerifyUrl) {
-  const verifyUrlRequest = fetch(`${environment.API_URL}/v0/sessions/identity_proof`, {
-    method: 'GET',
-    headers: new Headers({
-      Authorization: `Token token=${sessionStorage.userToken}`
-    })
-  }).then(response => {
-    return response.json();
-  }).then(json => {
-    if (json.identity_proof_url) {
-      onUpdateVerifyUrl(json.identity_proof_url);
-    }
-  });
-
-  return verifyUrlRequest;
-}
-
-export function getMultifactorUrl(onUpdateMultifactorUrl) {
-  const getMultifactorUrlRequest = fetch(`${environment.API_URL}/v0/sessions/multifactor`, {
-    method: 'GET',
-    headers: new Headers({
-      Authorization: `Token token=${sessionStorage.userToken}`
-    })
-  }).then(response => {
-    return response.json();
-  }).then(json => {
-    onUpdateMultifactorUrl(json.multifactor_url);
-  });
-
-  return getMultifactorUrlRequest;
-}
 
 export function getUserData(dispatch) {
   fetch(`${environment.API_URL}/v0/user`, {
@@ -85,6 +35,7 @@ export function getUserData(dispatch) {
       authnContext: userData.authn_context,
       loa: userData.loa,
       multifactor: userData.multifactor,
+      verified: userData.verified,
       gender: userData.gender,
       dob: userData.birth_date,
       status: json.data.attributes.va_profile.status,
@@ -105,40 +56,4 @@ export function getUserData(dispatch) {
     }
     dispatch(profileLoadingFinished());
   });
-}
-
-export function addEvent(element, eventName, callback) {
-  if (element.addEventListener) {
-    element.addEventListener(eventName, callback, false);
-  } else if (element.attachEvent) {
-    element.attachEvent('on' + eventName, callback); // eslint-disable-line prefer-template
-  }
-}
-
-export function getLoginUrls(dispatch) {
-  const loginUrlsRequest = fetch(`${environment.API_URL}/v0/sessions/authn_urls`, {
-    method: 'GET',
-  }).then(response => {
-    if (response.ok) {
-      return response.json();
-    }
-    throw Error(response.statusText);
-  }).then(json => {
-    dispatch(updateLogInUrls(json));
-  }).catch(() => {
-    dispatch({ type: FETCH_LOGIN_URLS_FAILED });
-  });
-
-  return loginUrlsRequest;
-}
-
-export function handleLogin(loginUrl) {
-  window.dataLayer.push({ event: 'login-link-clicked' });
-  if (loginUrl) {
-    window.dataLayer.push({ event: 'login-link-opened' });
-    const receiver = window.open(`${loginUrl}&op=signin`, 'signinPopup', 'resizable=yes,scrollbars=1,top=50,left=500,width=500,height=750');
-    receiver.focus();
-  } else {
-    Promise.reject('Could not log in; loginUrl not provided.');
-  }
 }
