@@ -19,7 +19,7 @@ class SearchHelpSignIn extends React.Component {
     }
   }
 
-  handleSigninSignup = (e) => {
+  handleSignInSignUp = (e) => {
     e.preventDefault();
     window.dataLayer.push({ event: 'login-link-clicked' });
     this.props.toggleLoginModal(true);
@@ -50,12 +50,11 @@ class SearchHelpSignIn extends React.Component {
           this.props.toggleSearchHelpUserMenu('account', !login.utilitiesMenuIsOpen.account);
         }}
         greeting={greeting}
-        isOpen={login.utilitiesMenuIsOpen.account}
-        onUserLogout={this.props.onUserLogout}/>);
+        isOpen={login.utilitiesMenuIsOpen.account}/>);
     } else {
       const classes = classNames({ disabled: isLoading });
       content = (<div>
-        <a href="#" className={classes} onClick={this.handleSigninSignup}><span>Sign In</span><span className="signin-spacer">|</span><span>Sign Up</span></a>
+        <a href="#" className={classes} onClick={this.handleSignInSignUp}><span>Sign In</span><span className="signin-spacer">|</span><span>Sign Up</span></a>
       </div>
       );
     }
