@@ -32,11 +32,11 @@ describe('<ConfirmationPage>', () => {
       <ConfirmationPage form={form}/>
     );
 
-    expect(tree.subTree('.confirmation-page-title').text()).to.equal('Claim received');
+    expect(tree.subTree('.confirmation-page-title').text()).to.equal('Claim submitted');
     expect(tree.everySubTree('span')[1].text().trim()).to.equal('for Jane Doe');
     expect(tree.everySubTree('li')[2].text()).to.contain('Western Region');
     expect(tree.everySubTree('p')[0].text()).to.contain('We process claims in the order we receive them');
-    expect(tree.everySubTree('p')[1].text()).to.contain('We may contact you for more information or documents.Please print this page for your records');
+    expect(tree.everySubTree('p')[1].text()).to.contain('We may contact you for more information or documents.');
     expect(tree.everySubTree('p')[3].text()).to.contain('VA Regional Office');
   });
 });
